refactor: Restructure runner. Different runner types are placed in different modules
