+++ src/schemathesis/types.py
@@ -22,3 +22,5 @@ class NotSet:
 Filter = Union[str, List[str], Tuple[str], Set[str], NotSet]  # pragma: no mutate
 
 Hook = Callable[[SearchStrategy], SearchStrategy]  # pragma: no mutate
+
+RawAuth = Tuple[str, str]  # pragma: no mutate
